+++ packages/web-components/src/components/hash-router/components/hash-router.ts
@@ -2,8 +2,8 @@ import type { CorrectComponentType } from "../../../utils/solid-element";
 
 import html from "solid-js/html";
 import { HashRouter as CustomRouter } from "../router";
-import { getHashRouterContextId } from "../hooks/useRouter";
-import { provideRoutesContext } from "../hooks/useRoutes";
+import { getHashRouterContextId } from "../hooks/useRouter.js";
+import { provideRoutesContext } from "../hooks/useRoutes.js";
 
 const HashRouter: CorrectComponentType<{}> = (_, { element }) => {
   Object.assign(element, { routerContextId: getHashRouterContextId() });