+++ packages/web-components/src/components/tabs/components/tab-items.ts
@@ -1,6 +1,6 @@
 import type { CorrectComponentType } from "../../../utils/solid-element";
 
-import { provideTabItemsContext } from "../hooks/useTabItems";
+import { provideTabItemsContext } from "../hooks/useTabItems.js";
 
 type TabItemsProps = {
   activeTab: string;