+++ packages/web-components/src/hooks/index.ts
@@ -1,1 +1,1 @@
-export { default as createQueueable } from "./createQueable";
+export { default as createQueueable } from "./createQueable.js";